Fixed ModuleAngleFromJoystick.java
I was doing the dead band wrong.

To prove what I had before was wrong, analyze when x=1 and y=0. x is more than 0.3 so it skips to the else. In the else, I was subtracting both numbers by 0.3, but y=0 so we don't want to do that because -0.3 when the joystick says 0.0 doesn't make sense.

So instead of remapping x and y, I remapped the velocity so 0.3 on the joystick is 0 velocity and 1.0 on the joystick is 1.0 velocity (or whatever MAX_SPEED is set to).