Improved NT values for Rotation
Previously, the rotError was pulled from poseError which is not accurate because that's from the Trajectory however due to the way SwerveControllerCommand is setup that Rotation is not the actual setpoint.

So I changed it to grab the proper rotation setpoint and to grab the heading from the gyro and subtract those to get the rotation error. Also added a new NT entry for the rotation setpoint.